rosteriq/requirements.txt
streamlit
pandas
duckdb
langchain
langchain-openai
pyyaml
plotly
tavily-python
python-dotenv
